refactor: プロジェクト読み込みを check.rs から scenario/project.rs に抽出する
trace（#77）が check と同じ読み込み規則（.md 閉包・README 除外・相対パス解決）を
再利用するための Tidy First。挙動変更なし（テスト 21 件 + 単体テスト緑を確認）。
