+++ src/scenario/check.rs
@@ -16,9 +16,11 @@
 
 use super::characters::{Characters, find_characters_file, load_characters};
 use super::diagnostic::{Diagnostic, Severity, Span};
-use super::parse::{Parsed, parse_file};
-use super::{Block, LinkTarget, Scene, slugify};
-use std::collections::{BTreeMap, HashMap, HashSet, VecDeque};
+use super::project::{
+    LoadedScene, collect_md_files, file_level, load_project, resolve_sibling, scene_links,
+};
+use super::{Block, LinkTarget, slugify};
+use std::collections::{BTreeMap, HashMap, HashSet};
 use std::path::{Path, PathBuf};
 
 /// check の動作オプション
@@ -60,15 +62,6 @@ impl CheckResult {
     pub fn has_errors(&self) -> bool {
         self.error_count() > 0
     }
-}
-
-/// 読み込んだ 1 シーン
-struct LoadedScene {
-    /// 表示用パス（入力引数からの相対のまま Diagnostic に載せる）
-    path: PathBuf,
-    /// 同一ファイル判定用の正規化パス
-    canon: PathBuf,
-    parsed: Parsed,
 }
 
 /// Markdown ファイルまたはディレクトリを検査する（SPEC 6章）。
@@ -142,119 +135,6 @@ pub fn check_path(path: &Path, options: &CheckOptions) -> CheckResult {
         files: scenes.iter().map(|s| s.path.clone()).collect(),
         diagnostics,
     }
-}
-
-// ---------------------------------------------------------------- 読み込み
-
-/// ディレクトリ配下の `.md` を再帰的に集める（名前順）。
-/// 隠しディレクトリと、シーンではない README.md は除く（SPEC 6章）
-fn collect_md_files(dir: &Path, out: &mut Vec<PathBuf>) {
-    let Ok(entries) = std::fs::read_dir(dir) else {
-        return;
-    };
-    let mut entries: Vec<_> = entries.flatten().collect();
-    entries.sort_by_key(|e| e.file_name());
-    for entry in entries {
-        let name = entry.file_name().to_string_lossy().to_string();
-        if name.starts_with('.') || name.eq_ignore_ascii_case("readme.md") {
-            continue;
-        }
-        let path = entry.path();
-        if path.is_dir() {
-            collect_md_files(&path, out);
-        } else if path.extension().and_then(|e| e.to_str()) == Some("md") {
-            out.push(path);
-        }
-    }
-}
-
-/// seeds とそこからリンクで辿れる `.md` をすべてパースする
-fn load_project(seeds: Vec<PathBuf>, diagnostics: &mut Vec<Diagnostic>) -> Vec<LoadedScene> {
-    let mut scenes = Vec::new();
-    let mut seen: HashSet<PathBuf> = HashSet::new();
-    let mut queue: VecDeque<PathBuf> = seeds.into();
-    while let Some(display) = queue.pop_front() {
-        // 実在は enqueue 前に確認済みだが、競合で消えた場合は黙って飛ばさず報告する
-        let canon = match display.canonicalize() {
-            Ok(c) => c,
-            Err(e) => {
-                diagnostics.push(file_level(
-                    "io-error",
-                    Severity::Error,
-                    &display,
-                    format!("{} を読み込めません: {}", display.display(), e),
-                ));
-                continue;
-            }
-        };
-        if !seen.insert(canon.clone()) {
-            continue;
-        }
-        let parsed = match parse_file(&display) {
-            Ok(p) => p,
-            Err(e) => {
-                diagnostics.push(file_level("io-error", Severity::Error, &display, e));
-                continue;
-            }
-        };
-        // リンク先の .md も検査対象に加える（閉包）。実在しない・絶対パスの
-        // ファイルは check_links が broken-link として報告する
-        for (_, target, _) in scene_links(&parsed.scene) {
-            if let Some(file) = &target.file
-                && let Some(resolved) = resolve_sibling(&display, file)
-                && resolved.is_file()
-                && resolved.extension().and_then(|e| e.to_str()) == Some("md")
-            {
-                queue.push_back(resolved);
-            }
-        }
-        scenes.push(LoadedScene {
-            path: display,
-            canon,
-            parsed,
-        });
-    }
-    scenes
-}
-
-/// シーンファイルからの相対パスを解決する。
-/// tsumugai が読むのは相対パスで参照されるファイルだけ（SPEC 2章）なので、
-/// 絶対パスは解決せず None を返す（呼び出し側が broken-link / missing-asset に倒す）
-fn resolve_sibling(scene_path: &Path, relative: &str) -> Option<PathBuf> {
-    if Path::new(relative).is_absolute() {
-        return None;
-    }
-    Some(
-        scene_path
-            .parent()
-            .unwrap_or_else(|| Path::new(""))
-            .join(relative),
-    )
-}
-
-/// シーン内のすべてのリンク（ジャンプ + 選択肢項目）を (ラベル, 飛び先, 行) で列挙する
-fn scene_links(scene: &Scene) -> Vec<(&str, &LinkTarget, usize)> {
-    let mut out = Vec::new();
-    let blocks = scene
-        .lead
-        .iter()
-        .chain(scene.sections.iter().flat_map(|s| s.blocks.iter()));
-    for block in blocks {
-        match block {
-            Block::Jump {
-                label,
-                target,
-                line,
-            } => out.push((label.as_str(), target, *line)),
-            Block::Choices { items, .. } => {
-                for item in items {
-                    out.push((item.label.as_str(), &item.target, item.line));
-                }
-            }
-            _ => {}
-        }
-    }
-    out
 }
 
 // ------------------------------------------------------- duplicate-scene-id
@@ -726,24 +606,6 @@ fn block_line(block: &Block) -> usize {
         | Block::Choices { line, .. }
         | Block::Jump { line, .. }
         | Block::Ending { line, .. } => *line,
-    }
-}
-
-/// span を持たないファイルレベルの Diagnostic
-fn file_level(
-    rule_id: &'static str,
-    severity: Severity,
-    file: &Path,
-    message: String,
-) -> Diagnostic {
-    Diagnostic {
-        rule_id,
-        severity,
-        message,
-        file: file.to_path_buf(),
-        span: None,
-        related_spans: Vec::new(),
-        suggestion: None,
     }
 }
